# Improve encoding generation performance by not duplicating data obj