--- a/package.json
+++ b/package.json
@@ -16,6 +16,8 @@
     "test": "test"
   },
   "scripts": {
+    "build": "gulp build",
+    "watch": "gulp build watch",
     "deploy": "npm run lint && npm run test && scripts/deploy.sh",
     "lint": "gulp jshint",
     "test": "gulp t"