@kadamwhite 2015-10-14
This does not fully resolve #120, because JSON serialization/deserialization is still inefficient and leads to more garbage collection than a smarter algorithm like the [structured clone](http://www.w3.org/TR/html5/infrastructure.html#internal-structured-cloning-algorithm) approach taken by [lodash.clonedeep](https://www.npmjs.com/package/lodash.clonedeep): A complete solution to #120 would possibly entail both identifying whether any of this duplication is strictly not-necessary, but also investigating whether `lodash.clonedeep` should be pulled in as a dev dependency of `datalib` as a replacement for the existing `duplicate` method.