@arvind 2015-10-14
@kadamwhite -- I'm really enjoying lurking on these PRs as I'm learning a lot about how to profile JavaScript applications! Would you consider writing a tutorial? :)