@kadamwhite 2015-10-14
@arvind Thanks, I'm glad it's been interesting! I'd love to write a piece for the Bocoup blog, but I'm not sure when I'll have time. There's a lot of great resources out there on both profiling applications and on specific Angular performance concerns already, although I don't have a specific one I could recommend at present.