@kadamwhite 2015-10-14
@kanitw @domoritz this is ready for review