@kanitw 2015-10-16
> This does not fully resolve #120, because JSON serialization/deserialization is still inefficient and leads to more garbage collection than a smarter algorithm like the [structured clone](http://www.w3.org/TR/html5/infrastructure.html#internal-structured-cloning-algorithm) approach taken by [lodash.clonedeep](https://www.npmjs.com/package/lodash.clonedeep): A complete solution to #120 would possibly entail both identifying whether any of this duplication is strictly not-necessary, but also investigating whether `lodash.clonedeep` should be pulled in as a dev dependency of `datalib` as a replacement for the existing `duplicate` method.

#120 says that "duplication is slow with large data" and this issue has sufficient solves. 

Actually serialization/deserialization is (at least "was") the fastest way to clone javascript object (without function) in __many browsers__ as the serialization/deserialization code is written in C++ or some fast language, so it's faster than `_.cloneDeep`.   (See discussion on http://stackoverflow.com/questions/122102/what-is-the-most-efficient-way-to-clone-an-object)

If you still feel that the clone method is better implemented in other ways across browsers, please file an issue on datalib.