@kadamwhite 2015-10-16
@kanitw Interesting reference! I was curious about whether that was still the case, so I made a [JSPerf test](http://jsperf.com/serialization-vs-deep-clone) to pit `JSON.parse/JSON.serialize` against `_.cloneDeep`. Even 7 years after that stack overflow issue, the initial results suggest that stringification may indeed still be faster:

![image](https://cloud.githubusercontent.com/assets/442115/10533601/7edd1c8e-7399-11e5-813a-e310c213b310.png)

However,  I was also concerned by the memory and garbage collection implications, which I haven't been able to conclusively test (I have a possibly-unfounded concern that stringification will incur a higher GC burden): and either way, the basic statement that we should avoid deep cloning objects (and deep watchers) wherever possible definitely remains true :)

Thanks for prompting me to investigate the issue further!